refactor: Move function from format module to component module
